+++ ds_dbw_can/include/ds_dbw_can/PlatformVersion.hpp
@@ -51,6 +51,7 @@ enum class Platform : uint8_t {
   FORD_P702        = 0x07, // Ford F150 (2021+)
   FORD_V3          = 0x08, // Ford E-Transit
   FORD_P702R       = 0x09, // Ford F150 Raptor (2024+)
+  FORD_NGDV        = 0x0A, // Ford/Oshkosh NGDV
   FCA_RU           = 0x10, // Chrysler Pacifica
   FCA_WK2          = 0x11, // Jeep Grand Cherokee
   POLARIS_GEM      = 0x80, // Polaris GEM
@@ -82,6 +83,7 @@ constexpr static const char* platformToString(Platform x) {
     case Platform::FORD_P702:        return "FORD_P702";
     case Platform::FORD_V3:          return "FORD_V3";
     case Platform::FORD_P702R:       return "FORD_P702R";
+    case Platform::FORD_NGDV:        return "FORD_NGDV";
     case Platform::FCA_RU:           return "FCA_RU";
     case Platform::FCA_WK2:          return "FCA_WK2";
     case Platform::POLARIS_GEM:      return "POLARIS_GEM";