+++ ds_dbw_can/src/DbwNode.cpp
@@ -87,45 +87,49 @@ static can_msgs::msg::Frame FrameFromDbw(const T &msg) {
 
 // Latest firmware versions
 static const PlatformMap FIRMWARE_LATEST({
-  {PlatformVersion(Platform::FCA_RU,            Module::Gateway,  ModuleVersion(2,0,1))},
-  {PlatformVersion(Platform::FCA_RU,            Module::Brake,    ModuleVersion(2,0,1))},
-  {PlatformVersion(Platform::FCA_RU,            Module::Throttle, ModuleVersion(2,0,1))},
-  {PlatformVersion(Platform::FCA_RU,            Module::Shift,    ModuleVersion(2,0,1))},
-  {PlatformVersion(Platform::FORD_CD4,          Module::Gateway,  ModuleVersion(3,3,2))},
-  {PlatformVersion(Platform::FORD_CD4,          Module::Brake,    ModuleVersion(3,3,2))},
-  {PlatformVersion(Platform::FORD_CD4,          Module::Throttle, ModuleVersion(3,3,2))},
-  {PlatformVersion(Platform::FORD_CD4,          Module::Shift,    ModuleVersion(3,3,2))},
-  {PlatformVersion(Platform::FORD_CD5,          Module::Gateway,  ModuleVersion(2,1,2))},
-  {PlatformVersion(Platform::FORD_CD5,          Module::Throttle, ModuleVersion(2,1,2))},
-  {PlatformVersion(Platform::FORD_CD5,          Module::BOO,      ModuleVersion(2,1,2))},
-  {PlatformVersion(Platform::FORD_GE1,          Module::Gateway,  ModuleVersion(2,3,2))},
-  {PlatformVersion(Platform::FORD_GE1,          Module::Throttle, ModuleVersion(2,3,2))},
-  {PlatformVersion(Platform::FORD_GE1,          Module::Shift,    ModuleVersion(2,3,2))},
-  {PlatformVersion(Platform::FORD_GE1,          Module::Monitor,  ModuleVersion(2,3,2))},
-  {PlatformVersion(Platform::FORD_P702,         Module::Gateway,  ModuleVersion(2,3,2))},
-  {PlatformVersion(Platform::FORD_P702,         Module::Throttle, ModuleVersion(2,3,2))},
-  {PlatformVersion(Platform::FORD_P702,         Module::Shift,    ModuleVersion(2,3,2))},
-  {PlatformVersion(Platform::FORD_P702R,        Module::Gateway,  ModuleVersion(1,0,1))},
-  {PlatformVersion(Platform::FORD_P702R,        Module::Throttle, ModuleVersion(1,0,1))},
-  {PlatformVersion(Platform::FORD_P702R,        Module::Shift,    ModuleVersion(1,0,1))},
-  {PlatformVersion(Platform::FORD_U6,           Module::Gateway,  ModuleVersion(2,3,2))},
-  {PlatformVersion(Platform::FORD_U6,           Module::Brake,    ModuleVersion(2,3,2))},
-  {PlatformVersion(Platform::FORD_U6,           Module::Throttle, ModuleVersion(2,3,2))},
-  {PlatformVersion(Platform::FORD_U6,           Module::Shift,    ModuleVersion(2,3,2))},
-  {PlatformVersion(Platform::FORD_U6,           Module::BOO,      ModuleVersion(2,3,2))},
-  {PlatformVersion(Platform::FORD_V3,           Module::Gateway,  ModuleVersion(1,1,2))},
-  {PlatformVersion(Platform::FORD_V3,           Module::Throttle, ModuleVersion(1,1,2))},
-  {PlatformVersion(Platform::FORD_V3,           Module::Shift,    ModuleVersion(1,1,2))},
-  {PlatformVersion(Platform::FORD_V3,           Module::BOO,      ModuleVersion(1,1,2))},
-  {PlatformVersion(Platform::POLARIS_RANGERXD,  Module::Gateway,  ModuleVersion(1,3,0))},
-  {PlatformVersion(Platform::POLARIS_RANGERXD,  Module::Throttle, ModuleVersion(1,3,0))},
-  {PlatformVersion(Platform::POLARIS_RANGERXP,  Module::Gateway,  ModuleVersion(1,3,0))},
-  {PlatformVersion(Platform::POLARIS_RANGERXP,  Module::Throttle, ModuleVersion(1,3,0))},
-  {PlatformVersion(Platform::POLARIS_RZRR,      Module::Gateway,  ModuleVersion(1,3,0))},
-  {PlatformVersion(Platform::POLARIS_RZRR,      Module::Steer,    ModuleVersion(1,3,0))},
-  {PlatformVersion(Platform::POLARIS_RZRR,      Module::Throttle, ModuleVersion(1,3,0))},
-  {PlatformVersion(Platform::POLARIS_RZRXP,     Module::Gateway,  ModuleVersion(1,3,0))},
-  {PlatformVersion(Platform::POLARIS_RZRXP,     Module::Throttle, ModuleVersion(1,3,0))},
+  {PlatformVersion(Platform::FCA_RU,            Module::Gateway,  ModuleVersion(2,0,3))},
+  {PlatformVersion(Platform::FCA_RU,            Module::Brake,    ModuleVersion(2,0,3))},
+  {PlatformVersion(Platform::FCA_RU,            Module::Throttle, ModuleVersion(2,0,3))},
+  {PlatformVersion(Platform::FCA_RU,            Module::Shift,    ModuleVersion(2,0,3))},
+  {PlatformVersion(Platform::FORD_CD4,          Module::Gateway,  ModuleVersion(3,3,4))},
+  {PlatformVersion(Platform::FORD_CD4,          Module::Brake,    ModuleVersion(3,3,4))},
+  {PlatformVersion(Platform::FORD_CD4,          Module::Throttle, ModuleVersion(3,3,4))},
+  {PlatformVersion(Platform::FORD_CD4,          Module::Shift,    ModuleVersion(3,3,4))},
+  {PlatformVersion(Platform::FORD_CD5,          Module::Gateway,  ModuleVersion(2,1,4))},
+  {PlatformVersion(Platform::FORD_CD5,          Module::Throttle, ModuleVersion(2,1,4))},
+  {PlatformVersion(Platform::FORD_CD5,          Module::BOO,      ModuleVersion(2,1,4))},
+  {PlatformVersion(Platform::FORD_GE1,          Module::Gateway,  ModuleVersion(2,3,4))},
+  {PlatformVersion(Platform::FORD_GE1,          Module::Throttle, ModuleVersion(2,3,4))},
+  {PlatformVersion(Platform::FORD_GE1,          Module::Shift,    ModuleVersion(2,3,4))},
+  {PlatformVersion(Platform::FORD_GE1,          Module::Monitor,  ModuleVersion(2,3,4))},
+  {PlatformVersion(Platform::FORD_NGDV,         Module::Gateway,  ModuleVersion(1,0,0))},
+  {PlatformVersion(Platform::FORD_NGDV,         Module::Throttle, ModuleVersion(1,0,0))},
+  {PlatformVersion(Platform::FORD_NGDV,         Module::Shift,    ModuleVersion(1,0,0))},
+  {PlatformVersion(Platform::FORD_NGDV,         Module::BOO,      ModuleVersion(1,0,0))},
+  {PlatformVersion(Platform::FORD_P702,         Module::Gateway,  ModuleVersion(2,3,4))},
+  {PlatformVersion(Platform::FORD_P702,         Module::Throttle, ModuleVersion(2,3,4))},
+  {PlatformVersion(Platform::FORD_P702,         Module::Shift,    ModuleVersion(2,3,4))},
+  {PlatformVersion(Platform::FORD_P702R,        Module::Gateway,  ModuleVersion(1,0,3))},
+  {PlatformVersion(Platform::FORD_P702R,        Module::Throttle, ModuleVersion(1,0,3))},
+  {PlatformVersion(Platform::FORD_P702R,        Module::Shift,    ModuleVersion(1,0,3))},
+  {PlatformVersion(Platform::FORD_U6,           Module::Gateway,  ModuleVersion(2,3,4))},
+  {PlatformVersion(Platform::FORD_U6,           Module::Brake,    ModuleVersion(2,3,4))},
+  {PlatformVersion(Platform::FORD_U6,           Module::Throttle, ModuleVersion(2,3,4))},
+  {PlatformVersion(Platform::FORD_U6,           Module::Shift,    ModuleVersion(2,3,4))},
+  {PlatformVersion(Platform::FORD_U6,           Module::BOO,      ModuleVersion(2,3,4))},
+  {PlatformVersion(Platform::FORD_V3,           Module::Gateway,  ModuleVersion(1,1,4))},
+  {PlatformVersion(Platform::FORD_V3,           Module::Throttle, ModuleVersion(1,1,4))},
+  {PlatformVersion(Platform::FORD_V3,           Module::Shift,    ModuleVersion(1,1,4))},
+  {PlatformVersion(Platform::FORD_V3,           Module::BOO,      ModuleVersion(1,1,4))},
+  {PlatformVersion(Platform::POLARIS_RANGERXD,  Module::Gateway,  ModuleVersion(1,3,2))},
+  {PlatformVersion(Platform::POLARIS_RANGERXD,  Module::Throttle, ModuleVersion(1,3,2))},
+  {PlatformVersion(Platform::POLARIS_RANGERXP,  Module::Gateway,  ModuleVersion(1,3,2))},
+  {PlatformVersion(Platform::POLARIS_RANGERXP,  Module::Throttle, ModuleVersion(1,3,2))},
+  {PlatformVersion(Platform::POLARIS_RZRR,      Module::Gateway,  ModuleVersion(1,3,2))},
+  {PlatformVersion(Platform::POLARIS_RZRR,      Module::Steer,    ModuleVersion(1,3,2))},
+  {PlatformVersion(Platform::POLARIS_RZRR,      Module::Throttle, ModuleVersion(1,3,2))},
+  {PlatformVersion(Platform::POLARIS_RZRXP,     Module::Gateway,  ModuleVersion(1,3,2))},
+  {PlatformVersion(Platform::POLARIS_RZRXP,     Module::Throttle, ModuleVersion(1,3,2))},
 });
 
 using std::placeholders::_1;
@@ -143,9 +147,6 @@ DbwNode::DbwNode(const rclcpp::NodeOptions &options)
 
   // Frame ID
   frame_id_ = declare_parameter<std::string>("frame_id", frame_id_);
-
-  // Use system enable/disable buttons
-  buttons_ = declare_parameter<bool>("buttons", buttons_);
 
   // Warn on received CRC errors, commands, and unknown messages
   warn_crc_ = declare_parameter<bool>("warn_crc", warn_crc_);
@@ -171,11 +172,13 @@ DbwNode::DbwNode(const rclcpp::NodeOptions &options)
   pub_brake_info_ = create_publisher<ds_dbw_msgs::msg::BrakeInfo>("brake/info", 2);
   pub_propulsion_info_ = create_publisher<ds_dbw_msgs::msg::PropulsionInfo>("propulsion/info", 2);
   pub_steer_offset_ = create_publisher<ds_dbw_msgs::msg::SteeringOffset>("steering/offset", 2);
+  pub_remote_rpt_ = create_publisher<ds_dbw_msgs::msg::RemoteReport>("remote/report", 2);
   pub_ulc_ = create_publisher<ds_dbw_msgs::msg::UlcReport>("ulc/report", 2);
   pub_wheel_speeds_ = create_publisher<ds_dbw_msgs::msg::WheelSpeeds>("wheel_speeds", 2);
   pub_wheel_positions_ = create_publisher<ds_dbw_msgs::msg::WheelPositions>("wheel_positions", 2);
   pub_turn_signal_ = create_publisher<ds_dbw_msgs::msg::TurnSignalReport>("turn_signal/report", 2);
   pub_misc_ = create_publisher<ds_dbw_msgs::msg::MiscReport>("misc/report", 2);
+  pub_gpio_ = create_publisher<ds_dbw_msgs::msg::GpioReport>("gpio/report", 2);
   pub_drive_mode_ = create_publisher<ds_dbw_msgs::msg::DriveModeReport>("drive_mode/report", 2);
   pub_driver_assist_ = create_publisher<ds_dbw_msgs::msg::DriverAssist>("driver_assist", 2);
   pub_battery_ = create_publisher<ds_dbw_msgs::msg::Battery>("battery", 2);
@@ -205,6 +208,7 @@ DbwNode::DbwNode(const rclcpp::NodeOptions &options)
   sub_turn_signal_ = create_subscription<ds_dbw_msgs::msg::TurnSignalCmd>("turn_signal/cmd", 1, std::bind(&DbwNode::recvTurnSignalCmd, this, _1));
   sub_drive_mode_ = create_subscription<ds_dbw_msgs::msg::DriveModeCmd>("drive_mode/cmd", 1, std::bind(&DbwNode::recvDriveModeCmd, this, _1));
   sub_misc_ = create_subscription<ds_dbw_msgs::msg::MiscCmd>("misc/cmd", 1, std::bind(&DbwNode::recvMiscCmd, this, _1));
+  sub_gpio_ = create_subscription<ds_dbw_msgs::msg::GpioCmd>("gpio/cmd", 1, std::bind(&DbwNode::recvGpioCmd, this, _1));
   sub_ulc_ = create_subscription<ds_dbw_msgs::msg::UlcCmd>("ulc/cmd", 1, std::bind(&DbwNode::recvUlcCmd, this, _1));
   sub_monitor_cmd_ = create_subscription<ds_dbw_msgs::msg::MonitorCmd>("monitor/cmd", 1, std::bind(&DbwNode::recvMonitorCmd, this, _1));
   sub_calibrate_steering_ = create_subscription<std_msgs::msg::Empty>("steering/calibrate", 1, std::bind(&DbwNode::recvSteeringCalibrate, this, _1));
@@ -286,10 +290,12 @@ void DbwNode::recvCAN(const can_msgs::msg::Frame::ConstSharedPtr msg_can) {
             if (msg_steer_rpt_2_.valid(stamp)) {
               const auto &msg2 = msg_steer_rpt_2_.msg();
               out.degraded = msg2.degraded;
+              out.actuator_temp = msg2.actuatorTempDegC();
               out.limit_rate = msg2.getLimitRateDegS();
               out.limit_value = msg2.getLimitValueDeg();
               out.cmd_src.value = (uint8_t)msg2.cmd_src;
             } else {
+              out.actuator_temp = NAN;
               out.limit_rate = NAN;
               out.limit_value = NAN;
               out.cmd_src.value = (uint8_t)CmdSrc::User;
@@ -545,9 +551,11 @@ void DbwNode::recvCAN(const can_msgs::msg::Frame::ConstSharedPtr msg_can) {
             if (msg_gear_rpt_2_.valid(stamp)) {
               const auto &msg2 = msg_gear_rpt_2_.msg();
               out.degraded = msg2.degraded;
+              out.actuator_temp = msg2.actuatorTempDegC();
               out.req_brake_cal = msg2.req_brake_cal;
               out.cmd_src.value = (uint8_t)msg2.cmd_src;
             } else {
+              out.actuator_temp = NAN;
               out.cmd_src.value = (uint8_t)CmdSrc::User;
             }
             pub_gear_rpt_->publish(out);
@@ -742,6 +750,7 @@ void DbwNode::recvCAN(const can_msgs::msg::Frame::ConstSharedPtr msg_can) {
             if (msg.fault_gear_override == MonitorFault) { RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Monitor fault: Gear override calculation mismatch with DBW"); }
             if (msg.fault_gear_cmd == MonitorFault) { RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Monitor fault: Gear actuator command mismatch with OEM and DBW"); }
             if (msg.fault_gear_cmd_ulc == MonitorFault) { RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Monitor fault: Gear command generated by ULC command without matching ULC command"); }
+            if (msg.fault_system_param == MonitorFault) { RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Monitor fault: System parameter mismatch with DBW"); }
           } else if (!recv.validCrc()) {
             RCLCPP_WARN(get_logger(), "Ignoring monitor report 3 with invalid CRC");
           } else if (!recv.validRc()) {
@@ -818,7 +827,7 @@ void DbwNode::recvCAN(const can_msgs::msg::Frame::ConstSharedPtr msg_can) {
                 msg_thrtl_cmd_clear_ = true;
                 msg_ulc_cmd_clear_ = true;
               }
-            } else if (buttons_) {
+            } else {
               if (msg.btn_enable) {
                 enableSystem();
               } else if (msg.btn_disable) {
@@ -875,8 +884,8 @@ void DbwNode::recvCAN(const can_msgs::msg::Frame::ConstSharedPtr msg_can) {
             out.accel_pedal_qf.value = (uint8_t)msg.accel_pedal_qf;
             out.one_pedal.value = (uint8_t)msg.one_pedal_drive;
             out.engine_rpm = msg.engineRpm();
+            out.engine_throttle_valve_pc = msg.engineThrottleValvePercent();
             out.drive_mode.value = (uint8_t)msg.drive_mode;
-            out.gear_num.value = (uint8_t)msg.gear_num;
             pub_thrtl_info_->publish(out);
           } else if (!recv.validCrc()) {
             RCLCPP_WARN(get_logger(), "Ignoring throttle info with invalid CRC");
@@ -922,6 +931,26 @@ void DbwNode::recvCAN(const can_msgs::msg::Frame::ConstSharedPtr msg_can) {
           auto &recv = msg_propulsion_info_;
           if (recv.receive(*(MsgPropulsionInfo*)msg_can->data.data(), stamp)) {
             const auto &msg = recv.msg();
+            static_assert(ds_dbw_msgs::msg::GearNum::UNKNOWN  == (uint8_t)MsgPropulsionInfo::GearNumber::Unknown);
+            static_assert(ds_dbw_msgs::msg::GearNum::DRIVE01  == (uint8_t)MsgPropulsionInfo::GearNumber::Drive01);
+            static_assert(ds_dbw_msgs::msg::GearNum::DRIVE02  == (uint8_t)MsgPropulsionInfo::GearNumber::Drive02);
+            static_assert(ds_dbw_msgs::msg::GearNum::DRIVE03  == (uint8_t)MsgPropulsionInfo::GearNumber::Drive03);
+            static_assert(ds_dbw_msgs::msg::GearNum::DRIVE04  == (uint8_t)MsgPropulsionInfo::GearNumber::Drive04);
+            static_assert(ds_dbw_msgs::msg::GearNum::DRIVE05  == (uint8_t)MsgPropulsionInfo::GearNumber::Drive05);
+            static_assert(ds_dbw_msgs::msg::GearNum::DRIVE06  == (uint8_t)MsgPropulsionInfo::GearNumber::Drive06);
+            static_assert(ds_dbw_msgs::msg::GearNum::DRIVE07  == (uint8_t)MsgPropulsionInfo::GearNumber::Drive07);
+            static_assert(ds_dbw_msgs::msg::GearNum::DRIVE08  == (uint8_t)MsgPropulsionInfo::GearNumber::Drive08);
+            static_assert(ds_dbw_msgs::msg::GearNum::DRIVE09  == (uint8_t)MsgPropulsionInfo::GearNumber::Drive09);
+            static_assert(ds_dbw_msgs::msg::GearNum::DRIVE10  == (uint8_t)MsgPropulsionInfo::GearNumber::Drive10);
+            static_assert(ds_dbw_msgs::msg::GearNum::REVERSE1 == (uint8_t)MsgPropulsionInfo::GearNumber::Reverse1);
+            static_assert(ds_dbw_msgs::msg::GearNum::REVERSE2 == (uint8_t)MsgPropulsionInfo::GearNumber::Reverse2);
+            static_assert(ds_dbw_msgs::msg::GearNum::NEUTRAL  == (uint8_t)MsgPropulsionInfo::GearNumber::Neutral);
+            static_assert(ds_dbw_msgs::msg::GearNum::PARK     == (uint8_t)MsgPropulsionInfo::GearNumber::Park);
+            static_assert(ds_dbw_msgs::msg::Bool::UNKNOWN == (uint8_t)MsgPropulsionInfo::Bool::Unknown);
+            static_assert(ds_dbw_msgs::msg::Bool::FALSE   == (uint8_t)MsgPropulsionInfo::Bool::False);
+            static_assert(ds_dbw_msgs::msg::Bool::TRUE    == (uint8_t)MsgPropulsionInfo::Bool::True);
+            static_assert(ds_dbw_msgs::msg::Bool::ERROR   == (uint8_t)MsgPropulsionInfo::Bool::Error);
+
             ds_dbw_msgs::msg::PropulsionInfo out;
             out.header.stamp = msg_can->header.stamp;
             out.propulsion_torque_request = msg.propulsionTorqueRequest();
@@ -929,6 +958,14 @@ void DbwNode::recvCAN(const can_msgs::msg::Frame::ConstSharedPtr msg_can) {
             out.engine_torque_request = msg.engineTorqueRequest();
             out.engine_torque_actual = msg.engineTorqueActual();
             out.transmission_ratio = msg.transmissionRatio();
+            if (msg.oem == MsgPropulsionInfo::Oem::Ford) {
+              out.gear_num_target.value = (uint8_t)msg.ford.gear_num_target;
+              out.gear_num_actual.value = (uint8_t)msg.ford.gear_num_actual;
+            }
+            if (msg.oem == MsgPropulsionInfo::Oem::Polaris) {
+              out.torque_limiting.value = (uint8_t)msg.polaris.torque_limiting;
+              out.limp_home.value = (uint8_t)msg.polaris.limp_home;
+            }
             pub_propulsion_info_->publish(out);
           } else if (!recv.validCrc()) {
             RCLCPP_WARN(get_logger(), "Ignoring propulsion info with invalid CRC");
@@ -959,6 +996,32 @@ void DbwNode::recvCAN(const can_msgs::msg::Frame::ConstSharedPtr msg_can) {
           }
         } else {
           RCLCPP_WARN(get_logger(), "Ignoring steer offset with invalid size of %u", msg_can->dlc);
+        }
+        break;
+
+      case MsgRemoteReport::ID:
+        if (msg_can->dlc >= sizeof(MsgRemoteReport)) {
+          auto &recv = msg_remote_rpt_;
+          auto remote_control_mode_prev = recv.msg().remote_control_mode;
+          if (recv.receive(*(MsgRemoteReport*)msg_can->data.data(), stamp)) {
+            const auto &msg = recv.msg();
+            ds_dbw_msgs::msg::RemoteReport out;
+            out.header.stamp = msg_can->header.stamp;
+            out.remote_control_mode.value = (uint8_t)msg.remote_control_mode;
+            out.remote_control_state.value = (uint8_t)msg.remote_control_state;
+            out.timeout = msg.timeout;
+            pub_remote_rpt_->publish(out);
+            if (remote_control_mode_prev != msg.remote_control_mode || !remote_control_mode_printed_) {
+              remote_control_mode_printed_ = true;
+              RCLCPP_INFO(get_logger(), "Remote control mode: %s", MsgRemoteReport::modeToString(msg.remote_control_mode));
+            }
+          } else if (!recv.validCrc()) {
+            RCLCPP_WARN(get_logger(), "Ignoring remote report with invalid CRC");
+          } else if (!recv.validRc()) {
+            RCLCPP_WARN(get_logger(), "Ignoring remote report with repeated rolling counter value");
+          }
+        } else {
+          RCLCPP_WARN(get_logger(), "Ignoring remote report with invalid size of %u", msg_can->dlc);
         }
         break;
 
@@ -1172,6 +1235,89 @@ void DbwNode::recvCAN(const can_msgs::msg::Frame::ConstSharedPtr msg_can) {
           }
         } else {
           RCLCPP_WARN(get_logger(), "Ignoring misc report 2 report with invalid size of %u", msg_can->dlc);
+        }
+        break;
+
+      case MsgMiscReport3::ID:
+        if (msg_can->dlc == sizeof(MsgMiscReport3)) {
+          auto &recv = msg_misc_rpt_3_;
+          if (recv.receive(*(MsgMiscReport3*)msg_can->data.data(), stamp)) {
+            const auto &msg = recv.msg();
+            if (msg.engine_coolant_hot_warn) {
+              RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Engine coolant over-temperature");
+            }
+            if (msg.engine_oil_hot_warn) {
+              RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Engine oil over-temperature");
+            }
+            if (msg.trans_oil_hot_warn) {
+              RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Transmission oil over-temperature");
+            }
+            if (msg.trans_clutch_hot_warn) {
+              RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Transmission clutch over-temperature");
+            }
+            if (msg.traction_motor_coil_hot_warn) {
+              RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Traction motor winding over-temperature");
+            }
+            if (msg.traction_motor_elec_hot_warn) {
+              RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Traction motor controller over-temperature");
+            }
+          } else if (!recv.validCrc()) {
+            RCLCPP_WARN(get_logger(), "Ignoring misc report 3 with invalid CRC");
+          } else if (!recv.validRc()) {
+            RCLCPP_WARN(get_logger(), "Ignoring misc report 3 with repeated rolling counter value");
+          }
+        } else {
+          RCLCPP_WARN(get_logger(), "Ignoring misc report 3 report with invalid size of %u", msg_can->dlc);
+        }
+        break;
+
+      case MsgGpioReport::ID:
+        if (msg_can->dlc == sizeof(MsgGpioReport)) {
+          auto &recv = msg_gpio_rpt_;
+          if (recv.receive(*(MsgGpioReport*)msg_can->data.data(), stamp)) {
+            const auto &msg = recv.msg();
+            static_assert(ds_dbw_msgs::msg::GpioReport::CMD_PASSIVE == (uint8_t)MsgGpioReport::GpioCmd::Passive);
+            static_assert(ds_dbw_msgs::msg::GpioReport::CMD_LOW     == (uint8_t)MsgGpioReport::GpioCmd::Low);
+            static_assert(ds_dbw_msgs::msg::GpioReport::CMD_PASSIVE == (uint8_t)MsgGpioReport::SsrCmd::Passive);
+            static_assert(ds_dbw_msgs::msg::GpioReport::CMD_ACTIVE  == (uint8_t)MsgGpioReport::SsrCmd::Active);
+            static_assert(ds_dbw_msgs::msg::GpioReport::GPIO_LOW  == (uint8_t)MsgGpioReport::GpioStatus::Low);
+            static_assert(ds_dbw_msgs::msg::GpioReport::GPIO_HIGH == (uint8_t)MsgGpioReport::GpioStatus::High);
+
+            ds_dbw_msgs::msg::GpioReport out;
+            out.header.stamp = msg_can->header.stamp;
+            out.gpio1_cmd = (uint8_t)msg.gpio1_cmd;
+            out.gpio2_cmd = (uint8_t)msg.gpio2_cmd;
+            out.gpio3_cmd = (uint8_t)msg.gpio3_cmd;
+            out.gpio4_cmd = (uint8_t)msg.gpio4_cmd;
+            out.gpio5_cmd = (uint8_t)msg.gpio5_cmd;
+            out.gpio6_cmd = (uint8_t)msg.gpio6_cmd;
+            out.ssr1_cmd = (uint8_t)msg.ssr1_cmd;
+            out.ssr2_cmd = (uint8_t)msg.ssr2_cmd;
+            out.gpio1_status = (uint8_t)msg.gpio1_status;
+            out.gpio2_status = (uint8_t)msg.gpio2_status;
+            out.gpio3_status = (uint8_t)msg.gpio3_status;
+            out.gpio4_status = (uint8_t)msg.gpio4_status;
+            out.gpio5_status = (uint8_t)msg.gpio5_status;
+            out.gpio6_status = (uint8_t)msg.gpio6_status;
+            out.gpio1_cmd_supported = msg.gpio1_cmd_supported;
+            out.gpio2_cmd_supported = msg.gpio2_cmd_supported;
+            out.gpio3_cmd_supported = msg.gpio3_cmd_supported;
+            out.gpio4_cmd_supported = msg.gpio4_cmd_supported;
+            out.gpio5_cmd_supported = msg.gpio5_cmd_supported;
+            out.gpio6_cmd_supported = msg.gpio6_cmd_supported;
+            out.ssr1_cmd_supported = msg.ssr1_cmd_supported;
+            out.ssr2_cmd_supported = msg.ssr2_cmd_supported;
+            out.timeout = msg.timeout;
+            out.bad_crc = msg.bad_crc;
+            out.bad_rc = msg.bad_rc;
+            pub_gpio_->publish(out);
+          } else if (!recv.validCrc()) {
+            RCLCPP_WARN(get_logger(), "Ignoring GPIO report with invalid CRC");
+          } else if (!recv.validRc()) {
+            RCLCPP_WARN(get_logger(), "Ignoring GPIO report with repeated rolling counter value");
+          }
+        } else {
+          RCLCPP_WARN(get_logger(), "Ignoring GPIO report with invalid size of %u", msg_can->dlc);
         }
         break;
 
@@ -1436,10 +1582,12 @@ void DbwNode::recvCAN(const can_msgs::msg::Frame::ConstSharedPtr msg_can) {
               out.degraded_comms_dbw_brake = msg3.degraded_comms_dbw_brake;
               out.degraded_comms_dbw_thrtl = msg3.degraded_comms_dbw_thrtl;
               out.degraded_comms_dbw_gear = msg3.degraded_comms_dbw_gear;
+              out.degraded_stall = msg3.degraded_stall;
               out.degraded_control_performance = msg3.degraded_control_performance;
               out.degraded_param_mismatch = msg3.degraded_param_mismatch;
               out.degraded_comms_vehicle = msg3.degraded_comms_vehicle;
               out.degraded_comms_actuator = msg3.degraded_comms_actuator;
+              out.degraded_actuator_warm = msg3.degraded_actuator_warm;
               out.degraded_vehicle_speed = msg3.degraded_vehicle_speed;
               out.degraded_calibration = msg3.degraded_calibration;
               out.fault_comms_dbw = msg3.fault_comms_dbw;
@@ -1457,6 +1605,7 @@ void DbwNode::recvCAN(const can_msgs::msg::Frame::ConstSharedPtr msg_can) {
               out.fault_actuator_torque_sensor = msg3.fault_actuator_torque_sensor;
               out.fault_actuator_config = msg3.fault_actuator_config;
               out.fault_actuator_assist = msg3.fault_actuator_assist;
+              out.fault_actuator_hot = msg3.fault_actuator_hot;
               out.fault_control_performance = msg3.fault_control_performance;
               out.fault_param_mismatch = msg3.fault_param_mismatch;
               out.fault_param_limits = msg3.fault_param_limits;
@@ -1559,6 +1708,7 @@ void DbwNode::recvCAN(const can_msgs::msg::Frame::ConstSharedPtr msg_can) {
               out.degraded_bped_feedback = msg3.degraded_bped_feedback;
               out.degraded_btsi_stuck_low = msg3.degraded_btsi_stuck_low;
               out.degraded_btsi_stuck_high = msg3.degraded_btsi_stuck_high;
+              out.degraded_hold_duration = msg3.degraded_hold_duration;
               out.degraded_actuator_aeb_deny = msg3.degraded_actuator_aeb_deny;
               out.degraded_actuator_1 = msg3.degraded_actuator_1;
               out.degraded_actuator_2 = msg3.degraded_actuator_2;
@@ -1581,6 +1731,8 @@ void DbwNode::recvCAN(const can_msgs::msg::Frame::ConstSharedPtr msg_can) {
               out.fault_bped_sensor_mismatch = msg3.fault_bped_sensor_mismatch;
               out.fault_actuator_1 = msg3.fault_actuator_1;
               out.fault_actuator_2 = msg3.fault_actuator_2;
+              out.fault_actuator_hold_duration = msg3.fault_actuator_hold_duration;
+              out.fault_actuator_hot = msg3.fault_actuator_hot;
               out.fault_control_performance = msg3.fault_control_performance;
               out.fault_param_mismatch = msg3.fault_param_mismatch;
               out.fault_param_limits = msg3.fault_param_limits;
@@ -1815,6 +1967,7 @@ void DbwNode::recvCAN(const can_msgs::msg::Frame::ConstSharedPtr msg_can) {
               out.degraded_comms_actuator = msg3.degraded_comms_actuator;
               out.degraded_comms_actuator_1 = msg3.degraded_comms_actuator_1;
               out.degraded_comms_actuator_2 = msg3.degraded_comms_actuator_2;
+              out.degraded_actuator_warm = msg3.degraded_actuator_warm;
               out.degraded_vehicle_speed = msg3.degraded_vehicle_speed;
               out.degraded_gear_mismatch = msg3.degraded_gear_mismatch;
               out.degraded_stuck_in_neutral = msg3.degraded_stuck_in_neutral;
@@ -1831,6 +1984,7 @@ void DbwNode::recvCAN(const can_msgs::msg::Frame::ConstSharedPtr msg_can) {
               out.fault_comms_actuator = msg3.fault_comms_actuator;
               out.fault_comms_actuator_1 = msg3.fault_comms_actuator_1;
               out.fault_comms_actuator_2 = msg3.fault_comms_actuator_2;
+              out.fault_actuator_hot = msg3.fault_actuator_hot;
               out.fault_vehicle_speed = msg3.fault_vehicle_speed;
               out.fault_actuator_config = msg3.fault_actuator_config;
               out.fault_param_mismatch = msg3.fault_param_mismatch;
@@ -1902,6 +2056,9 @@ void DbwNode::recvCAN(const can_msgs::msg::Frame::ConstSharedPtr msg_can) {
             if (msg.degraded_comms_dbw_gear) {
               RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Steering degraded: Lost comms with other Drive-By-Wire gear module");
             }
+            if (msg.degraded_stall) {
+              RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Steering degraded: Stall detected and limiting active");
+            }
             if (msg.degraded_control_performance) {
               RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Steering degraded: Insufficient control performance");
             }
@@ -1913,6 +2070,9 @@ void DbwNode::recvCAN(const can_msgs::msg::Frame::ConstSharedPtr msg_can) {
             }
             if (msg.degraded_comms_actuator) {
               RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Steering degraded: Lost comms with actuator");
+            }
+            if (msg.degraded_actuator_warm) {
+              RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Steering degraded: Actuator warm");
             }
             if (msg.degraded_vehicle_speed) {
               RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Steering degraded: Unknown or invalid vehicle speed");
@@ -1966,6 +2126,9 @@ void DbwNode::recvCAN(const can_msgs::msg::Frame::ConstSharedPtr msg_can) {
             if (msg.fault_actuator_assist) {
               RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Steering fault: Torque assist unavailable");
             }
+            if (msg.fault_actuator_hot) {
+              RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Steering fault: Actuator hot");
+            }
             if (msg.fault_control_performance) {
               RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Steering fault: Insufficient control performance");
             }
@@ -1991,7 +2154,7 @@ void DbwNode::recvCAN(const can_msgs::msg::Frame::ConstSharedPtr msg_can) {
                       || firmware_.get(Platform::POLARIS_RANGERXD, Module::Gateway).valid()
                       || firmware_.get(Platform::POLARIS_RANGERXP, Module::Gateway).valid()
                       || firmware_.get(Platform::POLARIS_RZRXP,    Module::Gateway).valid()) {
-                txt = "Steering calibration fault. Press the two steering multiplier buttons at the same "
+                txt = "Steering calibration fault. Press the two steering multiplier buttons (back/select) at the same "
                       "time to set the center offset when the wheel is straight.";
               }
               RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "%s", txt);
@@ -2053,6 +2216,9 @@ void DbwNode::recvCAN(const can_msgs::msg::Frame::ConstSharedPtr msg_can) {
             }
             if (msg.degraded_btsi_stuck_high) {
               RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Brake degraded: BTSI stuck high");
+            }
+            if (msg.degraded_hold_duration) {
+              RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Brake degraded: Brake hold duration exceeded");
             }
             if (msg.degraded_actuator_aeb_deny) {
               RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Brake degraded: AEB deny from actuator");
@@ -2119,6 +2285,12 @@ void DbwNode::recvCAN(const can_msgs::msg::Frame::ConstSharedPtr msg_can) {
             }
             if (msg.fault_actuator_2) {
               RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Brake fault: Fault in actuator 2");
+            }
+            if (msg.fault_actuator_hold_duration) {
+              RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Brake fault: Actuator brake hold duration exceeded");
+            }
+            if (msg.fault_actuator_hot) {
+              RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Brake fault: Actuator hot");
             }
             if (msg.fault_control_performance) {
               RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Brake fault: Insufficient control performance");
@@ -2272,6 +2444,9 @@ void DbwNode::recvCAN(const can_msgs::msg::Frame::ConstSharedPtr msg_can) {
             if (msg.degraded_comms_actuator_2) {
               RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Gear degraded: Lost secondary comms with actuator");
             }
+            if (msg.degraded_actuator_warm) {
+              RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Gear degraded: Actuator warm");
+            }
             if (msg.degraded_vehicle_speed) {
               RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Gear degraded: Unknown or invalid vehicle speed");
             }
@@ -2319,6 +2494,9 @@ void DbwNode::recvCAN(const can_msgs::msg::Frame::ConstSharedPtr msg_can) {
             }
             if (msg.fault_comms_actuator_2) {
               RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Gear fault: Lost secondary comms with actuator");
+            }
+            if (msg.fault_actuator_hot) {
+              RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Gear fault: Actuator hot");
             }
             if (msg.fault_vehicle_speed) {
               RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 10e3, "Gear fault: Unknown or invalid vehicle speed");
@@ -2519,6 +2697,17 @@ void DbwNode::recvCAN(const can_msgs::msg::Frame::ConstSharedPtr msg_can) {
           if (param_hash_.thrtl != msg.hash) {
             param_hash_.thrtl = msg.hash;
             RCLCPP_INFO(get_logger(), "Throttle param hash: %08X", msg.hash);
+          }
+        }
+        break;
+      case MsgSystemParamHash::ID:
+        if (msg_can->dlc >= sizeof(MsgSystemParamHash)) {
+          auto &recv = msg_system_param_hash_;
+          recv.receive(*(MsgSystemParamHash*)msg_can->data.data(), stamp);
+          const auto &msg = recv.msg();
+          if (param_hash_.system != msg.hash) {
+            param_hash_.system = msg.hash;
+            RCLCPP_INFO(get_logger(), "System param hash: %08X", msg.hash);
           }
         }
         break;
@@ -2872,6 +3061,7 @@ void DbwNode::recvCanMisc(const std::vector<can_msgs::msg::Frame::ConstSharedPtr
   assert(msgs[1]->dlc == sizeof(MsgMiscReport2));
   const MsgMiscReport1 &msg1 = *reinterpret_cast<const MsgMiscReport1 *>(msgs[0]->data.data());
   const MsgMiscReport2 &msg2 = *reinterpret_cast<const MsgMiscReport2 *>(msgs[1]->data.data());
+  auto stamp = ros_clock_.now();
   ds_dbw_msgs::msg::MiscReport out;
   out.header.stamp = msgs[0]->header.stamp;
   out.parking_brake.value = (uint8_t)msg1.parking_brake;
@@ -2927,18 +3117,40 @@ void DbwNode::recvCanMisc(const std::vector<can_msgs::msg::Frame::ConstSharedPtr
   out.headlight_high_control.value = (uint8_t)msg2.headlight_high_control;
   out.ambient_light.value = (uint8_t)msg2.ambient_light;
   out.outside_air_temp = msg2.outsideAirTempDegC();
+  if (msg_misc_rpt_3_.valid(stamp)) {
+    out.engine_coolant_temp = msg_misc_rpt_3_.msg().engineCoolantTempDegC();
+    out.engine_oil_temp = msg_misc_rpt_3_.msg().engineOilTempDegC();
+    out.trans_oil_temp = msg_misc_rpt_3_.msg().transOilTempDegC();
+    out.trans_clutch_temp = msg_misc_rpt_3_.msg().transClutchTempDegC();
+    out.traction_motor_coil_temp = msg_misc_rpt_3_.msg().tractionMotorCoilTempDegC();
+    out.traction_motor_elec_temp = msg_misc_rpt_3_.msg().tractionMotorElecTempDegC();
+    out.engine_coolant_hot_warn = msg_misc_rpt_3_.msg().engine_coolant_hot_warn;
+    out.engine_oil_hot_warn = msg_misc_rpt_3_.msg().engine_oil_hot_warn;
+    out.trans_oil_hot_warn = msg_misc_rpt_3_.msg().trans_oil_hot_warn;
+    out.trans_clutch_hot_warn = msg_misc_rpt_3_.msg().trans_clutch_hot_warn;
+    out.traction_motor_coil_hot_warn = msg_misc_rpt_3_.msg().traction_motor_coil_hot_warn;
+    out.traction_motor_elec_hot_warn = msg_misc_rpt_3_.msg().traction_motor_elec_hot_warn;
+  } else {
+    out.engine_coolant_temp = NAN;
+    out.engine_oil_temp = NAN;
+    out.trans_oil_temp = NAN;
+    out.trans_clutch_temp = NAN;
+    out.traction_motor_coil_temp = NAN;
+    out.traction_motor_elec_temp = NAN;
+  }
   pub_misc_->publish(out);
   printSyncDelta(msgs[0], msgs[1], "misc");
 }
 
 void DbwNode::recvSteeringCmd(const ds_dbw_msgs::msg::SteeringCmd::ConstSharedPtr msg) {
-  static_assert(ds_dbw_msgs::msg::SteeringCmd::CMD_NONE      == (uint8_t)MsgSteerCmd::CmdType::None);
-  static_assert(ds_dbw_msgs::msg::SteeringCmd::CMD_TORQUE    == (uint8_t)MsgSteerCmd::CmdType::Torque);
-  static_assert(ds_dbw_msgs::msg::SteeringCmd::CMD_ANGLE     == (uint8_t)MsgSteerCmd::CmdType::Angle);
-  static_assert(ds_dbw_msgs::msg::SteeringCmd::CMD_CURVATURE == (uint8_t)MsgSteerCmd::CmdType::Curvature);
-  static_assert(ds_dbw_msgs::msg::SteeringCmd::CMD_YAW_RATE  == (uint8_t)MsgSteerCmd::CmdType::YawRate);
-  static_assert(ds_dbw_msgs::msg::SteeringCmd::CMD_PERCENT   == (uint8_t)MsgSteerCmd::CmdType::Percent);
-  static_assert(ds_dbw_msgs::msg::SteeringCmd::CMD_CALIBRATE == (uint8_t)MsgSteerCmd::CmdType::Calibrate);
+  static_assert(ds_dbw_msgs::msg::SteeringCmd::CMD_NONE       == (uint8_t)MsgSteerCmd::CmdType::None);
+  static_assert(ds_dbw_msgs::msg::SteeringCmd::CMD_TORQUE     == (uint8_t)MsgSteerCmd::CmdType::Torque);
+  static_assert(ds_dbw_msgs::msg::SteeringCmd::CMD_ANGLE      == (uint8_t)MsgSteerCmd::CmdType::Angle);
+  static_assert(ds_dbw_msgs::msg::SteeringCmd::CMD_CURVATURE  == (uint8_t)MsgSteerCmd::CmdType::Curvature);
+  static_assert(ds_dbw_msgs::msg::SteeringCmd::CMD_YAW_RATE   == (uint8_t)MsgSteerCmd::CmdType::YawRate);
+  static_assert(ds_dbw_msgs::msg::SteeringCmd::CMD_PERCENT    == (uint8_t)MsgSteerCmd::CmdType::Percent);
+  static_assert(ds_dbw_msgs::msg::SteeringCmd::CMD_SET_OFFSET == (uint8_t)MsgSteerCmd::CmdType::SetOffset);
+  static_assert(ds_dbw_msgs::msg::SteeringCmd::CMD_CALIBRATE  == (uint8_t)MsgSteerCmd::CmdType::Calibrate);
   auto stamp = ros_clock_.now();
   if (std::isnan(msg->cmd) && msg->cmd_type != ds_dbw_msgs::msg::SteeringCmd::CMD_NONE) {
     RCLCPP_WARN(get_logger(), "NaN steering command");
@@ -2970,6 +3182,10 @@ void DbwNode::recvSteeringCmd(const ds_dbw_msgs::msg::SteeringCmd::ConstSharedPt
     case ds_dbw_msgs::msg::SteeringCmd::CMD_PERCENT:
       msg_steer_cmd_.cmd_type = MsgSteerCmd::CmdType::Percent;
       msg_steer_cmd_.setCmdPercentDeg(msg->cmd, msg->cmd_rate, msg->cmd_accel);
+      break;
+    case ds_dbw_msgs::msg::SteeringCmd::CMD_SET_OFFSET:
+      msg_steer_cmd_.cmd_type = MsgSteerCmd::CmdType::SetOffset;
+      msg_steer_cmd_.setCmdAngleDeg(msg->cmd);
       break;
     case ds_dbw_msgs::msg::SteeringCmd::CMD_CALIBRATE:
       msg_steer_cmd_.cmd_type = MsgSteerCmd::CmdType::Calibrate;
@@ -3262,6 +3478,22 @@ void DbwNode::recvMiscCmd(const ds_dbw_msgs::msg::MiscCmd::ConstSharedPtr msg) {
   pub_can_->publish(FrameFromDbw(msg_misc_cmd_));
 }
 
+void DbwNode::recvGpioCmd(const ds_dbw_msgs::msg::GpioCmd::ConstSharedPtr msg) {
+  msg_gpio_cmd_.reset();
+  msg_gpio_cmd_.gpio1_cmd = (MsgGpioCmd::GpioCmd)msg->gpio1_cmd;
+  msg_gpio_cmd_.gpio2_cmd = (MsgGpioCmd::GpioCmd)msg->gpio2_cmd;
+  msg_gpio_cmd_.gpio3_cmd = (MsgGpioCmd::GpioCmd)msg->gpio3_cmd;
+  msg_gpio_cmd_.gpio4_cmd = (MsgGpioCmd::GpioCmd)msg->gpio4_cmd;
+  msg_gpio_cmd_.gpio5_cmd = (MsgGpioCmd::GpioCmd)msg->gpio5_cmd;
+  msg_gpio_cmd_.gpio6_cmd = (MsgGpioCmd::GpioCmd)msg->gpio6_cmd;
+  msg_gpio_cmd_.ssr1_cmd = (MsgGpioCmd::SsrCmd)msg->ssr1_cmd;
+  msg_gpio_cmd_.ssr2_cmd = (MsgGpioCmd::SsrCmd)msg->ssr2_cmd;
+  msg_gpio_cmd_.rc++;
+  msg_gpio_cmd_.setCrc();
+
+  pub_can_->publish(FrameFromDbw(msg_gpio_cmd_));
+}
+
 void DbwNode::recvUlcCmd(const ds_dbw_msgs::msg::UlcCmd::ConstSharedPtr msg) {
   if (std::isnan(msg->cmd) && msg->cmd_type != ds_dbw_msgs::msg::UlcCmd::CMD_NONE) {
     RCLCPP_WARN(get_logger(), "NaN ULC command");
@@ -3359,13 +3591,14 @@ void DbwNode::recvMonitorCmd(const ds_dbw_msgs::msg::MonitorCmd::ConstSharedPtr 
 
 void DbwNode::recvSteeringCalibrate(const std_msgs::msg::Empty::ConstSharedPtr) {
   /* Send steering command to save current angle as zero.
-   * The preferred method is to set the 'calibrate' field in a ROS steering
-   * command so that recvSteeringCmd() saves the current angle as the
-   * specified command.
+   * The preferred method is to set the command type as 'calibrate' in a ROS
+   * steering command so that recvSteeringCmd() saves the current angle as the
+   * specified command, or set the command type as "set offset" in a ROS steering
+   * command so that recvSteeringCmd() saves the offset as the specified command.
    */
   RCLCPP_WARN_THROTTLE(get_logger(), *get_clock(), 1e3,
       "The std_msgs/Empty steering calibration topic is deprecated. "
-      "Instead, send a steering command with cmd_type=CMD_CALIBRATE and specify the angle"
+      "Instead, send a steering command with cmd_type CMD_CALIBRATE or CMD_SET_OFFSET and specify the angle"
   );
   msg_steer_cmd_.reset();
   msg_steer_cmd_.cmd_type = MsgSteerCmd::CmdType::Calibrate;